Enforces auth and creates company page.
Auth state is now synced across components using the useAuth hook and
enforced in the login and signup pages.

The api wil now automatically include accesToken in request header and
call the refreshToken endpoint when requests fail with HTTP 419 and
store and use the new tokens.

Adds a working create-connection page.